// src/types/form.type.ts
export interface Translation {
  name: string;
  surname: string;
  email: string;
  message: string;
  enterName: string;
  enterSurname: string;
  enterEmail: string;
  enterMessage: string;
}

export interface ButtonText {
  send: string;
}

export interface FormData {
  firstName: string;
  lastName: string;
  email: string;
  message: string;
}